requirements.txt
discord.py
dotenv
pytz
datetime
aiohttp